link error

+++ lab6/runtime.c
@@ -3,10 +3,9 @@
 #include <stdlib.h>
 #include <string.h>
 
-
 int *initArray(int size, int init)
 {int i;
- int *a = (int *)malloc(size*sizeof(int));
+ int *a = (int *)malloc(size*sizeof(long));
  for(i=0;i<size;i++) a[i]=init;
  return a;
 }
